refactor(draft): change name querymodel to readmodel
