+++ services/draft/application/src/main/java/nettee/draft/application/usecase/DraftReadByStatusesUseCase.java
@@ -1,6 +1,6 @@
 package nettee.draft.application.usecase;
 
-import nettee.draft.readmodel.DraftQueryModels.DraftSummary;
+import nettee.draft.readmodel.DraftReadModels.DraftSummary;
 import nettee.draft.domain.type.DraftStatus;
 import org.springframework.data.domain.Page;
