+++ services/draft/application/src/main/java/nettee/draft/application/usecase/DraftReadUseCase.java
@@ -1,7 +1,7 @@
 package nettee.draft.application.usecase;
 
-import nettee.draft.readmodel.DraftQueryModels.DraftDetail;
-import nettee.draft.readmodel.DraftQueryModels.DraftSummary;
+import nettee.draft.readmodel.DraftReadModels.DraftDetail;
+import nettee.draft.readmodel.DraftReadModels.DraftSummary;
 import org.springframework.data.domain.Page;
 
 import java.util.Optional;
