+++ services/draft/application/src/main/java/nettee/draft/draftblock/application/port/DraftBlockCommandPort.java
@@ -2,7 +2,7 @@ package nettee.draft.draftblock.application.port;
 
 import nettee.draft.draftblock.domain.DraftBlock;
 import nettee.draft.draftblock.domain.type.DraftBlockStatus;
-import nettee.draft.draftblock.readmodel.DraftBlockQueryModels.DraftBlockDetail;
+import nettee.draft.draftblock.readmodel.DraftBlockReadModels.DraftBlockDetail;
 
 import java.util.Optional;
