+++ services/draft/application/src/main/java/nettee/draft/draftblock/application/service/DraftBlockQueryService.java
@@ -5,8 +5,8 @@ import nettee.draft.draftblock.application.port.DraftBlockQueryPort;
 import nettee.draft.draftblock.application.usecase.DraftBlockReadByStatusesUseCase;
 import nettee.draft.draftblock.application.usecase.DraftBlockReadUseCase;
 import nettee.draft.draftblock.domain.type.DraftBlockStatus;
-import nettee.draft.draftblock.readmodel.DraftBlockQueryModels.DraftBlockDetail;
-import nettee.draft.draftblock.readmodel.DraftBlockQueryModels.DraftBlockSummary;
+import nettee.draft.draftblock.readmodel.DraftBlockReadModels.DraftBlockDetail;
+import nettee.draft.draftblock.readmodel.DraftBlockReadModels.DraftBlockSummary;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
