+++ services/draft/application/src/main/java/nettee/draft/draftblock/application/usecase/DraftBlockReadByStatusesUseCase.java
@@ -1,7 +1,7 @@
 package nettee.draft.draftblock.application.usecase;
 
 import nettee.draft.draftblock.domain.type.DraftBlockStatus;
-import nettee.draft.draftblock.readmodel.DraftBlockQueryModels.DraftBlockSummary;
+import nettee.draft.draftblock.readmodel.DraftBlockReadModels.DraftBlockSummary;
 import org.springframework.data.domain.Page;
 
 import java.util.Set;
