+++ services/draft/application/src/main/java/nettee/draft/draftblock/application/usecase/DraftBlockReadUseCase.java
@@ -1,7 +1,7 @@
 package nettee.draft.draftblock.application.usecase;
 
-import nettee.draft.draftblock.readmodel.DraftBlockQueryModels.DraftBlockDetail;
-import nettee.draft.draftblock.readmodel.DraftBlockQueryModels.DraftBlockSummary;
+import nettee.draft.draftblock.readmodel.DraftBlockReadModels.DraftBlockDetail;
+import nettee.draft.draftblock.readmodel.DraftBlockReadModels.DraftBlockSummary;
 import org.springframework.data.domain.Page;
 
 import java.util.Optional;
