+++ services/draft/driven/rdb/src/main/java/nettee/draft/driven/rdb/DraftCommandAdapter.java
@@ -1,7 +1,7 @@
 package nettee.draft.driven.rdb;
 
 import lombok.RequiredArgsConstructor;
-import nettee.draft.readmodel.DraftQueryModels.DraftDetail;
+import nettee.draft.readmodel.DraftReadModels.DraftDetail;
 import nettee.draft.domain.Draft;
 import nettee.draft.driven.rdb.entity.type.DraftEntityStatus;
 import nettee.draft.driven.rdb.persistence.mapper.DraftEntityMapper;
