+++ services/draft/driven/rdb/src/main/java/nettee/draft/driven/rdb/DraftQueryAdapter.java
@@ -1,7 +1,7 @@
 package nettee.draft.driven.rdb;
 
-import nettee.draft.readmodel.DraftQueryModels.DraftDetail;
-import nettee.draft.readmodel.DraftQueryModels.DraftSummary;
+import nettee.draft.readmodel.DraftReadModels.DraftDetail;
+import nettee.draft.readmodel.DraftReadModels.DraftSummary;
 import nettee.draft.driven.rdb.entity.DraftEntity;
 import nettee.draft.driven.rdb.entity.type.DraftEntityStatus;
 import nettee.draft.driven.rdb.persistence.mapper.DraftEntityMapper;
