+++ services/draft/driven/rdb/src/main/java/nettee/draft/driven/rdb/persistence/mapper/DraftEntityMapper.java
@@ -1,8 +1,8 @@
 package nettee.draft.driven.rdb.persistence.mapper;
 
 import nettee.draft.driven.rdb.entity.DraftEntity;
-import nettee.draft.readmodel.DraftQueryModels.DraftDetail;
-import nettee.draft.readmodel.DraftQueryModels.DraftSummary;
+import nettee.draft.readmodel.DraftReadModels.DraftDetail;
+import nettee.draft.readmodel.DraftReadModels.DraftSummary;
 import nettee.draft.domain.Draft;
 import org.mapstruct.Mapper;
