+++ services/draft/driving/web-mvc/src/main/java/nettee/draft/draftblock/driving/web/dto/DraftBlockQueryDto.java
@@ -1,8 +1,7 @@
 package nettee.draft.draftblock.driving.web.dto;
 
 import lombok.Builder;
-import nettee.draft.draftblock.readmodel.DraftBlockQueryModels.DraftBlockDetail;
-import nettee.draft.readmodel.DraftQueryModels.DraftDetail;
+import nettee.draft.draftblock.readmodel.DraftBlockReadModels.DraftBlockDetail;
 
 public class DraftBlockQueryDto {
     private DraftBlockQueryDto() {
