+++ services/draft/driving/web-mvc/src/main/java/nettee/draft/driving/web/dto/DraftQueryDto.java
@@ -1,7 +1,7 @@
 package nettee.draft.driving.web.dto;
 
 import lombok.Builder;
-import nettee.draft.readmodel.DraftQueryModels.DraftDetail;
+import nettee.draft.readmodel.DraftReadModels.DraftDetail;
 
 public class DraftQueryDto {
     private DraftQueryDto() {
