+++ services/draft/driving/web-mvc/src/main/java/nettee/draft/driving/web/mapper/DraftDtoMapper.java
@@ -1,7 +1,7 @@
 package nettee.draft.driving.web.mapper;
 
 import nettee.draft.domain.Draft;
-import nettee.draft.readmodel.DraftQueryModels.DraftDetail;
+import nettee.draft.readmodel.DraftReadModels.DraftDetail;
 import nettee.draft.driving.web.dto.DraftCommandDto.DraftCreateCommand;
 import nettee.draft.driving.web.dto.DraftCommandDto.DraftUpdateCommand;
 import nettee.draft.driving.web.dto.DraftCommandDto.DraftUpdateTotalLikesCommand;
